OBLS-833 Auto allocation job stops working when a failure occurs move functions to services
