+++ grails-app/services/org/pih/warehouse/receiving/ReceiptService.groovy
@@ -512,6 +512,22 @@ class ReceiptService {
         return null
     }
 
+    void receiveInboundShipments(Location facility) {
+        log.info "Detecting candidates for auto receipt - inbound shipments in-transit to facility ${facility}"
+        List<Shipment> shippedShipments = shipmentService.getShippedShipmentsByDestination(facility)
+        shippedShipments.each { Shipment shipment ->
+            try {
+                authService.withSystemUser {
+                    log.info "Creating automated receipt for shipment ${shipment}"
+                    receiveInboundShipment(shipment)
+                    reallocateBackorderedItems(shipment.id)
+                }
+            } catch (Exception e) {
+                log.error("Error processing requisition ${requisitionId}", e)
+            }
+        }
+    }
+
     void receiveInboundShipment(Shipment shipment, Boolean forceAutoReceipt = false) {
 
         log.info "Receive inbound shipment ${shipment}  forceAutoReceipt${forceAutoReceipt}"
